adding abstract persistence spec
targets Reader/Writer interfaces and leaves subclasses to implement the
instances
